+++ recipes/alertmanager_discovery.rb
@@ -1,6 +1,6 @@
 #
 # Cookbook:: rblx_prometheus
-# Recipe:: prom_discovery
+# Recipe:: alertmanager_discovery
 #
 # Copyright:: 2020, Roblox, All Rights Reserved.
 
@@ -9,29 +9,29 @@ infradb_available = (node.key?('infradb') and node['infradb'].key?('serverInfo')
 location_available = (infradb_available && !node['infradb']['serverInfo']['Server']['Location'].nil?)
 
 if node['rblx_prometheus']['config']['pod'].nil?
-  raise 'prom_discovery: pod unspecified by attributes and graphql unavailable' if !(location_available)
+  raise 'target_discovery: pod unspecified by attributes and graphql unavailable' if !(location_available)
   pod = node['infradb']['serverInfo']['Server']['Location']['Pod']['Name'].downcase()
 else
   pod = node['rblx_prometheus']['config']['pod']
 end
 
 if node['rblx_prometheus']['config']['datacenter'].nil?
-  raise 'prom_discovery: datacenter unspecified by attributes and graphql unavailable' if !(location_available)
+  raise 'target_discovery: datacenter unspecified by attributes and graphql unavailable' if !(location_available)
   dc = node['infradb']['serverInfo']['Server']['Location']['DataCenter']['Abbreviation'].downcase()
 else
   dc = node['rblx_prometheus']['config']['datacenter']
 end
 ##### end of copy-paste blob
 
-if node['rblx_prometheus']['config']['prometheus']['enable'] == true
+if node['rblx_prometheus']['config']['alertmanager']['enable'] == true
   require 'net/http'
   require 'uri'
   require 'json'
 
   addresses = []
-  prometheus_service = node['rblx_prometheus']['config']['prometheus']['consul_service']
+  alertmanager_service = node['rblx_prometheus']['config']['alertmanager']['consul_service']
   begin
-    uri = URI.parse("http://127.0.0.1:8500/v1/catalog/service/#{prometheus_service}")
+    uri = URI.parse("http://127.0.0.1:8500/v1/catalog/service/#{alertmanager_service}")
     response = Net::HTTP.get_response(uri)
   
     res = JSON.parse(response.body)
@@ -43,14 +43,14 @@ if node['rblx_prometheus']['config']['prometheus']['enable'] == true
     end if response.code
    
     if addresses.empty?
-      node.override['rblx_prometheus']['config']['prometheus']['_disable'] = true
-      node.override['rblx_prometheus']['config']['prometheus']['_target_list'] = ['EMPTY']
+      node.override['rblx_prometheus']['config']['alertmanager']['_disable'] = true
+      node.override['rblx_prometheus']['config']['alertmanager']['_target_list'] = ['EMPTY']
     else
-      node.override['rblx_prometheus']['config']['prometheus']['_target_list'] = addresses
+      node.override['rblx_prometheus']['config']['alertmanager']['_target_list'] = addresses
     end 
   rescue
     # We had some error looking up the consul service
-    node.override['rblx_prometheus']['config']['prometheus']['_disable'] = true
-    node.override['rblx_prometheus']['config']['prometheus']['_target_list'] = ['ERROR']
+    node.override['rblx_prometheus']['config']['alertmanager']['_disable'] = true
+    node.override['rblx_prometheus']['config']['alertmanager']['_target_list'] = ['ERROR']
   end
 end